# openapi.yaml
openapi: 3.1.0
info:
  title: CEOS ARD Editor API
  version: 0.1.0
  description: |
    RESTful API for the CEOS ARD Editor application that provides workspace management,
    file editing, building block management, and GitHub integration capabilities.
  contact:
    name: CEOS ARD Team
    url: https://github.com/ceos-org/ceos-ard-editor
  license:
    name: Apache-2.0
    url: https://github.com/ceos-org/ceos-ard-editor/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://ceos-ard-editor.example.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /workspaces:
    get:
      summary: List all workspaces
      description: Retrieves all workspaces for the authenticated user
      operationId: listWorkspaces
      tags:
        - Workspaces
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new workspace
      description: Creates a new workspace with cloned repository
      operationId: createWorkspace
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Workspace title
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  description: Workspace description
                pfs:
                  $ref: '#/components/schemas/pfs'
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /workspaces/{workspaceId}/previews:
    get:
      summary: Generate preview
      description: Generates preview for the current workspace state
      operationId: generatePreview
      tags:
        - Previews
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PfsType'
      responses:
        '200':
          description: Preview generated successfully
          content:
            text/html:
              schema:
                type: string
                description: HTML content of the preview
        '202':
          description: Preview generation in progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Preview generation started"
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/previews/{filePath}:
    get:
      summary: Return static assets for the preview
      description: Returns static assets for the preview of the current workspace state
      operationId: getPreviewAsset
      tags:
        - Previews
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PfsType'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: File content
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '304':
          description: Cached copy is still current; no body is returned.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/download:
    get:
      summary: Download preview document
      description: Downloads the preview document for the current workspace state
      operationId: downloadPreviewDocument
      tags:
        - Previews
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/PfsType'
        - $ref: '#/components/parameters/DocumentFormat'
      responses:
        '200':
          description: Preview document content downloaded successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Unique identifier of the workspace
      schema:
        type: string
        format: uuid
...
    PfsType:
      name: pfs
      in: query
      style: form
      explode: true
      description: |-
        PFS type for preview generation.

        If not provided, uses the default PFS type(s) that have been selected during workspace creation.
      schema:
        type: array
        items:
          type: string
    DocumentFormat:
      name: format
      in: query
      required: true
      description: Format of the downloaded preview document.
      schema:
        type: string
        enum:
          - pdf
          - docx
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Cached ETag to revalidate; if it still matches, the server returns 304 instead of the asset.
      schema:
        type: string
      example: '"5d41402abc4b2a76b9719d911017c592"'
  headers:
    ETag:
      description: Asset validator; send it in If-None-Match to revalidate.
      schema:
        type: string
      example: '"5d41402abc4b2a76b9719d911017c592"'
    LastModified:
      description: When the asset was last modified.
      schema:
        type: string
      example: 'Mon, 27 Jul 2026 10:00:00 GMT'
    CacheControl:
      description: Always no-cache; store but revalidate before reuse.
      schema:
        type: string
      example: 'no-cache'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
        - code
...
    Workspace:
      type: object
      required:
        - id
        - name
        - description
        - repository
        - branch
        - pfs
        - created
        - pr
        - user_id
      properties:
        id:
          type: string
          format: uuid
          description: Unique workspace identifier
        title:
          type: string
          description: Workspace title
        description:
          type: string
          description: Workspace description
        repository:
          type: string
          format: uri
          description: URL of the Git repository for the workspace
        branch:
          type: string
          description: Name of the branch in the repository
        pfs:
          $ref: '#/components/schemas/pfs'
        created:
          type: string
          format: date-time
          description: Workspace creation timestamp
        status:
          type: string
          enum:
            - active
            - archived
          description: |
            Workspace status.

            - "archived": Workspace has been archived (manually or due to merged/closed PR)
            - "active": Workspace is active and can be worked on
        archived_at:
          type:
            - string
            - "null"
          format: date-time
          description: |
            Timestamp when workspace was archived.

            Set automatically when status changes to "archived".
            Cleared when workspace is reactivated.
          default: null
        deletion_at:
          type:
            - string
            - "null"
          format: date-time
          description: |
            Scheduled permanent deletion timestamp (computed as archived_at + 1 month).

            **Note**: This is a computed property, not stored in the database.
            After this time, workspace files and database record will be permanently deleted by cleanup script.
            Automatically null when workspace is not archived.
          default: null
          readOnly: true
        pr:
          type:
            - string
            - "null"
          description: ID to the GitHub PR
          default: null
        pr_status:
          type:
            - string
            - "null"
          enum:
            - unknown
            - open
            - approved
            - closed
            - merged
            - null
          description: Status of the PR
          default: null
        user_id:
          type: string
          description: The user ID of the workspace owner
...
    pfs:
      description: List of PFS types assigned to the workspace, used for preview generation.
      type: array
      items:
        $ref: '#/components/schemas/pfsId'
      example:
        - NRB
        - POL
...
    pfsId:
      type: string
      description: PFS type identifier
      pattern: ^[A-Z0-9][A-Z0-9-]*[A-Z0-9]$
      example: NRB